[BYOB-151] 루트 경로 접속 시 /meetings 경로로 넘어가도록 리다이렉션 설정 (#26)
